refactor: update import paths in ProtectedRoute, Home components and adjust CSS file references
