+++ src/components/container/routing/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../../../context/AuthContext';
 
 const ProtectedRoute = ({ children }) => {
   const { user } = useSelector((state) => state.auth);
