+++ src/hooks/useAuth.js
@@ -1,5 +1,3 @@
-import { useSelector } from 'react-redux';
-
 export const useAuth = () => {
   const auth = useSelector((state) => state.auth);
