+++ src/pages/Home.jsx
@@ -1,6 +1,6 @@
 // Home.js
 import React from 'react';
-import './styles/Home.css';
+import '../styles/Home.css';
 import { Link } from 'react-router-dom';
 
 const Home = () => {
