+++ src/styles/Home.css
@@ -1,8 +1,15 @@
 .hero-section {
     background: linear-gradient(rgba(0,0,0,0.6), rgba(0,0,0,0.6)),
-                url('../../../public/images/tea-background.jpg');
+                url('../../public/images/tea-background.jpg');
     background-size: cover;
     background-position: center;
     height: 80vh;
     min-height: 500px;
   }
+
+
+
+
+
+
+
